refactor: 아이템 카테고리 구현 (#9)
